% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_hours_used.R
\name{app_hours_used}
\alias{app_hours_used}
\title{Total hours of app usage}
\usage{
app_hours_used(account = "scotland", back_months = 3)
}
\arguments{
\item{account}{Name of Shiny Server account where app is deployed.}

\item{back_months}{Number of months in the past to include app usage data.
Maximum value is 3.}
}
\description{
Total hours of app usage
}
\examples{
\dontrun{app_hours_used(back_months = 1)}
}
