% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_table.R
\name{app_table}
\alias{app_table}
\title{Create interactive table of apps}
\usage{
app_table(data)
}
\arguments{
\item{data}{Tibble of app data.}
}
\value{
An interactive \code{gt} table (of class \code{gt_tbl}).
}
\description{
Create interactive table of apps
}
