% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_visibility.R
\name{app_visibility}
\alias{app_visibility}
\title{Determine whether app is public or private}
\usage{
app_visibility(app_name, account = "scotland")
}
\arguments{
\item{app_name}{Character vector of names of apps on the Shiny Server.}

\item{account}{Name of Shiny Server account where app is deployed.}
}
\value{
Character vector of "public" or "private".
}
\description{
Determine whether app is public or private
}
\examples{
\dontrun{app_visibility("my_app")}
}
