% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_manual_url.R
\name{clean_manual_url}
\alias{clean_manual_url}
\title{Clean free-text app URLs}
\usage{
clean_manual_url(url, account = "scotland")
}
\arguments{
\item{url}{Character vector of untidy free-text URLs}

\item{account}{Name of Shiny Server account where app is deployed.}
}
\value{
Character vector of clean URLs
}
\description{
Clean free-text app URLs
}
\examples{
clean_manual_url("my_App")
}
