% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_output.R
\name{excel_output}
\alias{excel_output}
\title{Format app data into excel output}
\usage{
excel_output(analysis_table, org_group = "all")
}
\arguments{
\item{analysis_table}{Data frame of app data.}

\item{org_group}{Character. Either "all" to include all organisations in the
output, or an organisation group that appears in
\code{analysis_table$org_group}.}
}
\value{
Named list of data frames, that when using \code{writexl::write_xlsx},
will be sheets in an Excel file.
}
\description{
Format app data into excel output
}
