% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_org_prefix.R
\name{extract_org_prefix}
\alias{extract_org_prefix}
\title{Extract organisation prefix}
\usage{
extract_org_prefix(name)
}
\arguments{
\item{name}{Name of app (not URL)}
}
\value{
Character. Organisation prefix
}
\description{
Extract organisation prefix
}
\examples{
extract_org_prefix("sg-app")
}
