% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ms_form_names}
\alias{ms_form_names}
\title{Microsoft Form Names}
\format{
A tibble with 2 columns:
\describe{
\item{new}{Clean snakecase names to use in the analysis dataset}
\item{old}{Exact question wording in the MS form}
}
}
\usage{
ms_form_names
}
\description{
A lookup dataset for renaming columns in the Microsoft Forms contact dataset.
}
\keyword{datasets}
