% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_grouped_to_factor.R
\name{org_grouped_to_factor}
\alias{org_grouped_to_factor}
\title{Convert grouped organisation data to factor (for sorting)}
\usage{
org_grouped_to_factor(org_grouped)
}
\arguments{
\item{org_grouped}{Character vector of grouped organisations.
See Details for accepted values.}
}
\value{
Factor.
}
\description{
Convert grouped organisation data to factor (for sorting)
}
\details{
Accepted values in \code{org_grouped}:
\itemize{
\item Scottish Government (inc. agencies)
\item Public Health Scotland
\item Other
\item Unknown
}
}
