% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orgs}
\alias{orgs}
\title{Organisations}
\format{
A tibble with 4 columns:
\describe{
\item{org_acronym}{Organisation acronym used as prefix in app name}
\item{org_name}{Organisation name}
\item{sg_agency}{Logical. Is the organisation the Scottish Government or
one of its agencies?}
\item{ms_form_accepted}{Logical. Is the organisation a currently accepted
value in the Microsoft Form?}
}
}
\usage{
orgs
}
\description{
A lookup dataset for organisations with deployed apps.
}
\keyword{datasets}
