% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Progress bar for \code{purrr} functions}
\usage{
progress(name)
}
\arguments{
\item{name}{Progress bar name used in printed messages.}
}
\value{
List to be passed to \code{purrr} functions \code{.progress} argument.
}
\description{
Progress bar for \code{purrr} functions
}
\examples{
progress("Getting app visibility")
}
