% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_form_data.R
\name{refresh_form_data}
\alias{refresh_form_data}
\title{Refresh Microsoft Form data file}
\usage{
refresh_form_data(link, wait = 60)
}
\arguments{
\item{link}{Character string. Shared link to OneDrive data file containing
Microsoft Form responses.}

\item{wait}{Numeric. Number of seconds to wait to allow for local data file
to sync with OneDrive file. Defaults to 60.}
}
\value{
\code{TRUE} (invisibly)
}
\description{
The Microsoft Form to collect contact information is linked to
an Excel spreadsheet where results are stored. However, the spreadsheet
does not automatically update with new form responses (this is expected
behaviour for MS Forms).

This function is used to open the spreadsheet in a browser, allow time for
the online version to update and for the local version to sync.
}
\details{
If the value supplied to \code{link} doesn't exist or is an empty string,
the function will print a message and stop, but not produce an error.

The function does not check the validity of \code{link}. If invalid, the browser
will not open, but the function will continue to wait.

Running this function does not guarantee that the file will be refreshed.
The value provided to \code{wait} is arbitrary, and the user should make sure
this is long enough for the file to refresh in their working environment.
}
\examples{
refresh_form_data("x", wait = 10)
}
