% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols.R
\name{rename_cols}
\alias{rename_cols}
\title{Rename columns from lookup}
\usage{
rename_cols(df, names_lookup)
}
\arguments{
\item{df}{Data frame}

\item{names_lookup}{Data frame containing "new" and "old" columns.}
}
\value{
\code{df} with columns renamed according to \code{names_lookup}.
}
\description{
Rename columns from lookup
}
