% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_dashboard.R
\name{render_dashboard}
\alias{render_dashboard}
\title{Render quarto dashboard}
\usage{
render_dashboard(qmd_path, output_path)
}
\arguments{
\item{qmd_path}{Path to qmd script}

\item{output_path}{Path to desired output file destination}
}
\value{
\code{output_path} (invisibly).
}
\description{
Render quarto dashboard
}
