% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_apps.R
\name{server_apps}
\alias{server_apps}
\title{List deployed Shiny Server apps}
\usage{
server_apps(
  account = "scotland",
  visibility = FALSE,
  hours_used = FALSE,
  back_months = 3
)
}
\arguments{
\item{account}{Name of Shiny Server account where app is deployed.}

\item{visibility}{Logical: Should app visibility data be included?}

\item{hours_used}{Logical: Should hours used data be included?}

\item{back_months}{Number of months in the past to include app usage data.
Maximum value is 3.}
}
\value{
Tibble.
}
\description{
List deployed Shiny Server apps
}
\examples{
\dontrun{server_apps()}
}
