% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyadmin-package.R
\docType{package}
\name{shinyadmin-package}
\alias{shinyadmin}
\alias{shinyadmin-package}
\title{shinyadmin: Shiny App Server Administration}
\description{
Administration of the Scottish Public Sector Shiny Server (scotland.shinyapps.io).
}
\author{
\strong{Maintainer}: Alice Hannah \email{alice.hannah@gov.scot}

}
\keyword{internal}
