% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{start_app}
\alias{start_app}
\title{Launch test app}
\usage{
start_app(port = 8080, host = "0.0.0.0")
}
\arguments{
\item{port}{The port to run the app on}

\item{host}{The host to run the app on}
}
\description{
Launch the shiny app inst inst/app, for local testing
}
