% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_remove_dates.R
\name{url_remove_dates}
\alias{url_remove_dates}
\title{Remove dates from URLs}
\usage{
url_remove_dates(url)
}
\arguments{
\item{url}{Character vector of URLs}
}
\value{
Character vector
}
\description{
Remove dates from URLs
}
\examples{
url_remove_dates("https://scotland.shinyapps.io/my-app_feb25")
}
